refactor SubmitPanel into its own subdirectory
